+++ sdk/browser/src/AffinityWallet.ts
@@ -3,6 +3,8 @@ import { EventComponent } from '@affinidi/affinity-metrics-lib'
 
 import KeysService from './services/KeysService'
 import WalletStorageService from './services/WalletStorageService'
+import { FetchCredentialsPaginationOptions } from '@affinidi/wallet-core-sdk/dist/dto/shared.dto'
+import { MessageParameters } from '@affinidi/wallet-core-sdk/dist/dto'
 
 type SdkOptions = __dangerous.SdkOptions & {
   issueSignupCredential?: boolean
@@ -46,6 +48,34 @@ export class AffinityWallet extends CoreNetwork {
     const encryptionKey = await WalletStorageService.pullEncryptionKey(accessToken)
 
     return new AffinityWallet(encryptionKey, encryptedSeed, options)
+  }
+
+  /**
+   * @description Initiates sign up flow
+   * @param username - arbitrary username, email or phoneNumber
+   * @param password - is required if arbitrary username was provided.
+   * It is optional and random one will be generated, if not provided when
+   * email or phoneNumber was given as a username.
+   * @param options - optional parameters with specified environment
+   * @param messageParameters - optional parameters with specified welcome message
+   * @returns token or, in case when arbitrary username was used, it returns
+   * initialized instance of SDK
+   */
+  static async signUp(
+    username: string,
+    password?: string,
+    options?: SdkOptions,
+    messageParameters?: MessageParameters,
+  ): Promise<string | any> {
+    const networkMember = await CoreNetwork.signUp(username, password, options, messageParameters)
+
+    if (networkMember.constructor === String) {
+      return networkMember
+    }
+
+    const { password: networkMemberPassword, encryptedSeed } = networkMember
+
+    return new AffinityWallet(networkMemberPassword, encryptedSeed, options)
   }
 
   /**
@@ -176,19 +206,44 @@ export class AffinityWallet extends CoreNetwork {
   }
 
   /**
-   * @description Pulls all credentials which match by credentialShareRequestToken,
-   *   if token not provided all your VCs will be returned:
-   * 1. pull encrypted VCs
+   * @description Retrieve only the credential at given index
+   * @param credentialIndex - index for the VC in vault
+   * @returns a single VC
+   */
+  async getCredentialByIndex(credentialIndex: number): Promise<any> {
+    const paginationOptions: FetchCredentialsPaginationOptions = { skip: credentialIndex, limit: 1 }
+    const blobs = await this.walletStorageService.fetchEncryptedCredentials(paginationOptions)
+
+    if (blobs.length < 1) {
+      throw new __dangerous.SdkError('COR-14')
+    }
+
+    const decryptedCredentials = await this.walletStorageService.decryptCredentials(blobs)
+    return decryptedCredentials[0]
+  }
+
+  /**
+   * @description Finds the given credentialShareRequestToken by searching all of your credentials
+   * If a token is not given, only returns the given subset of the credentials
+   * 1. pull encrypted VCs (all if token given, otherwise with the given pagination)
    * 2. decrypt encrypted VCs
    * 3. filter VCs by type
    * @param credentialShareRequestToken - JWT received from verifier
+   * @param paginationOptions - optional range for credentials to be pulled (default is skip: 0, limit: 100)
    * @returns array of VCs
    */
-  async getCredentials(credentialShareRequestToken: string = null): Promise<any> {
+  async getCredentials(
+    credentialShareRequestToken: string = null,
+    paginationOptions?: FetchCredentialsPaginationOptions,
+  ): Promise<any> {
+    if (credentialShareRequestToken) {
+      return this._getCredentialsWithCredentialShareRequestToken(credentialShareRequestToken)
+    }
+
     let blobs
 
     try {
-      blobs = await this.walletStorageService.fetchEncryptedCredentials()
+      blobs = await this.walletStorageService.fetchEncryptedCredentials(paginationOptions)
     } catch (error) {
       if (error.code === 'COR-14') {
         return []
@@ -201,31 +256,46 @@ export class AffinityWallet extends CoreNetwork {
       return []
     }
 
-    const credentials = await this.walletStorageService.decryptCredentials(blobs)
+    return this.walletStorageService.decryptCredentials(blobs)
+  }
 
-    const encryptedCredentials = this.walletStorageService.filterCredentials(credentialShareRequestToken, credentials)
+  private async _getCredentialsWithCredentialShareRequestToken(credentialShareRequestToken: string): Promise<any> {
+    let result: any[] = []
 
-    return encryptedCredentials
+    for await (const blobs of this.walletStorageService.fetchAllEncryptedCredentialsInBatches()) {
+      const credentials = await this.walletStorageService.decryptCredentials(blobs)
+
+      const matchedCredentials = this.walletStorageService.filterCredentials(credentialShareRequestToken, credentials)
+
+      result = result.concat(matchedCredentials)
+    }
+
+    return result
   }
 
   /**
-   * @description Delete credential by id
+   * @description Delete credential by id if found in given range
    * @param id - id of the credential
+   * @param credentialIndex - credential to remove
    */
-  async deleteCredential(id: string): Promise<void> {
-    let blobs
-
-    try {
-      blobs = await this.walletStorageService.fetchEncryptedCredentials()
-    } catch (error) {
-      if (error.code === 'COR-14') {
-        throw new __dangerous.SdkError('COR-14')
-      } else {
-        throw error
-      }
+  async deleteCredential(id?: string, credentialIndex?: number): Promise<void> {
+    if ((credentialIndex !== undefined && id) || (!id && credentialIndex === undefined)) {
+      throw new __dangerous.SdkError('COR-1', {
+        errors: [{ message: 'should pass either id or credentialIndex and not both at the same time' }],
+      })
     }
 
-    await this.walletStorageService.decryptCredentials(blobs)
+    if (credentialIndex) {
+      return this.deleteCredentialByIndex(credentialIndex.toString())
+    }
+
+    let allBlobs: any[] = []
+
+    for await (const blobs of this.walletStorageService.fetchAllEncryptedCredentialsInBatches()) {
+      allBlobs = allBlobs.concat(blobs)
+    }
+
+    await this.walletStorageService.decryptCredentials(allBlobs)
 
     const credentialIndexToDelete = this.walletStorageService.findCredentialIndexById(id)