+++ sdk/core/test/unit/services/WalletStorageService.test.ts
@@ -10,6 +10,7 @@ import WalletStorageService from '../../../src/services/WalletStorageService'
 import { STAGING_VAULT_URL, STAGING_KEY_STORAGE_URL } from '../../../src/_defaultConfig'
 
 import { generateTestDIDs } from '../../factory/didFactory'
+import { authorizeVaultEndpoints } from '../../helpers/authorizeVault'
 
 const signedCredential = require('../../factory/signedCredential')
 
@@ -21,6 +22,7 @@ let walletPassword: string
 let encryptedSeed: string
 
 const fetchCredentialsPath = '/data/0/99'
+const fetchCredentialsBase = '/data/'
 
 describe('WalletStorageService', () => {
   before(async () => {
@@ -180,6 +182,111 @@ describe('WalletStorageService', () => {
     }
 
     expect(errorResponse.httpStatusCode).to.eql(500)
+  })
+
+  describe('#fetchAllEncryptedCredentialsInBatches', () => {
+    it('should work for single page', async () => {
+      await authorizeVault()
+
+      nock(STAGING_VAULT_URL).get(fetchCredentialsPath).reply(200, [signedCredential])
+
+      const walletStorageService = new WalletStorageService(encryptedSeed, walletPassword)
+      let response: any[] = []
+
+      for await (const blobs of walletStorageService.fetchAllEncryptedCredentialsInBatches()) {
+        response = response.concat(blobs)
+      }
+
+      expect(response).to.be.an('array')
+      expect(response).to.eql([signedCredential])
+    })
+
+    it('should work for multiple pages', async () => {
+      await authorizeVault()
+
+      nock(STAGING_VAULT_URL)
+        .get(fetchCredentialsBase + '0/0')
+        .reply(200, [signedCredential])
+
+      nock(STAGING_VAULT_URL)
+        .get(fetchCredentialsBase + '1/1')
+        .reply(200, [signedCredential])
+
+      nock(STAGING_VAULT_URL)
+        .get(fetchCredentialsBase + '2/2')
+        .reply(200, [])
+
+      const walletStorageService = new WalletStorageService(encryptedSeed, walletPassword)
+      let response: any[] = []
+
+      const paginationOptions = { skip: 0, limit: 1 }
+
+      for await (const blobs of walletStorageService.fetchAllEncryptedCredentialsInBatches(paginationOptions)) {
+        response = response.concat(blobs)
+
+        authorizeVaultEndpoints()
+      }
+
+      expect(response).to.be.an('array')
+      expect(response).to.eql([signedCredential, signedCredential])
+    })
+
+    it('should fail with invalid parameters', async () => {
+      const walletStorageService = new WalletStorageService(encryptedSeed, walletPassword)
+
+      let errorResponse
+
+      try {
+        const asyncIterable = walletStorageService.fetchAllEncryptedCredentialsInBatches({ skip: -1, limit: -1 })
+        const asyncIterator = asyncIterable[Symbol.asyncIterator]()
+
+        await asyncIterator.next()
+      } catch (error) {
+        errorResponse = error
+      }
+
+      expect(errorResponse).not.to.be.undefined
+      expect(errorResponse.code).to.be.eq('COR-1')
+    })
+  })
+
+  describe('#fetchEncryptedCredentials works with pagination', async () => {
+    const cases = [
+      { args: [{}], path: '0/99' },
+      { args: [{ limit: 5 }], path: '0/4' },
+      { args: [{ skip: 1 }], path: '1/100' },
+      { args: [{ skip: 1, limit: 1 }], path: '1/1' },
+    ]
+
+    for (const { args, path } of cases) {
+      it(`should request correct path with args ${JSON.stringify(args)}`, async () => {
+        await authorizeVault()
+
+        const expectedPath = fetchCredentialsBase + path
+        nock(STAGING_VAULT_URL).get(expectedPath).reply(200, [signedCredential])
+
+        const walletStorageService = new WalletStorageService(encryptedSeed, walletPassword)
+        const response = await walletStorageService.fetchEncryptedCredentials(...args)
+
+        expect(response).to.be.an('array')
+        expect(response).to.eql([signedCredential])
+      })
+    }
+
+    it('should fail with invalid parameters', async () => {
+      const walletStorageService = new WalletStorageService(encryptedSeed, walletPassword)
+
+      let errorResponse
+
+      try {
+        await walletStorageService.fetchEncryptedCredentials({ skip: -1, limit: -1 })
+      } catch (error) {
+        errorResponse = error
+      }
+
+      expect(errorResponse).not.to.be.undefined
+      expect(errorResponse.code).to.be.eq('COR-1')
+    })
   })
 
   it('#adminConfirmUser', async () => {